% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfsusieR-package.R
\docType{package}
\name{mfsusieR-package}
\alias{mfsusieR}
\alias{mfsusieR-package}
\title{mfsusieR: Multi-Modality Multi-Functional Sum of Single Effects Regression}
\description{
A multi-modality, multi-functional generalisation of the 'Sum of Single Effects' (SuSiE) model for variable selection in regression with mixtures of functional and univariate responses. Built on the 'susieR' backbone, with wavelet-based functional priors and a per-(scale, modality) variance structure.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StatFunGen/mfsusieR}
  \item Report bugs at \url{https://github.com/StatFunGen/mfsusieR/issues}
}

}
\author{
\strong{Maintainer}: Gao Wang \email{wang.gao@columbia.edu}

Authors:
\itemize{
  \item William R. P. Denault
  \item Anjing Liu
  \item Matthew Stephens
}

}
\keyword{internal}
